+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [1.3.0] - 18.04.2021
+
+* Fix with reinitilizing subscriptions for Events
+* Remove Android support
+
+## [1.2.0-nullsafety.0] - 12.03.2021
+
+* Null safety support
+
+## [1.1.1] - 12.03.2021
+
+* Anchored Query fix with deleted objects
+
 ## [1.1.0] - 25.02.2021
 
 * iOS 9.0 support